--- core-java-8/src/main/java/com/baeldung/datetime/DateToLocalDateTimeConverter.java
@@ -1,35 +1,0 @@
-/**
- * 
- */
-package com.baeldung.datetime;
-
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.util.Date;
-
-/**
- * Class which shows different ways of converting java.util.Date into java.time.LocalDateTime. 
- * 
- * @author abialas
- *
- */
-public class DateToLocalDateTimeConverter {
-
-    public static LocalDateTime convertToLocalDateTimeViaInstant(Date dateToConvert) {
-        return dateToConvert.toInstant()
-            .atZone(ZoneId.systemDefault())
-            .toLocalDateTime();
-    }
-
-    public static LocalDateTime convertToLocalDateTimeViaSqlTimestamp(Date dateToConvert) {
-        return new java.sql.Timestamp(dateToConvert.getTime()).toLocalDateTime();
-    }
-
-    public static LocalDateTime convertToLocalDateTimeViaMilisecond(Date dateToConvert) {
-        return Instant.ofEpochMilli(dateToConvert.getTime())
-            .atZone(ZoneId.systemDefault())
-            .toLocalDateTime();
-    }
-
-}
