+++ core-java-9/src/main/java/com/baeldung/java9/datetime/DateToLocalDateTimeConverter.java
@@ -3,17 +3,34 @@
  */
 package com.baeldung.java9.datetime;
 
+import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.Date;
 
 /**
- * Class which shows a way to convert java.util.Date into java.time.LocalDateTime with new Java 1.9.
+ * Class which shows a way to convert java.util.Date into java.time.LocalDateTime.
  * 
  * @author abialas
  *
  */
 public class DateToLocalDateTimeConverter {
+
+    public static LocalDateTime convertToLocalDateTimeViaInstant(Date dateToConvert) {
+        return dateToConvert.toInstant()
+            .atZone(ZoneId.systemDefault())
+            .toLocalDateTime();
+    }
+
+    public static LocalDateTime convertToLocalDateTimeViaSqlTimestamp(Date dateToConvert) {
+        return new java.sql.Timestamp(dateToConvert.getTime()).toLocalDateTime();
+    }
+
+    public static LocalDateTime convertToLocalDateTimeViaMilisecond(Date dateToConvert) {
+        return Instant.ofEpochMilli(dateToConvert.getTime())
+            .atZone(ZoneId.systemDefault())
+            .toLocalDateTime();
+    }
 
     public static LocalDateTime convertToLocalDateTime(Date dateToConvert) {
         return LocalDateTime.ofInstant(dateToConvert.toInstant(), ZoneId.systemDefault());
