+++ core-java-9/src/main/java/com/baeldung/java9/datetime/LocalDateTimeToDateConverter.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.baeldung.datetime;
+package com.baeldung.java9.datetime;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
