+++ core-java-9/src/test/java/com/baeldung/java9/datetime/LocalDateToDateConverterTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.baeldung.datetime;
+package com.baeldung.java9.datetime;
 
 import static org.junit.Assert.assertEquals;
